requirements.txt
# PDF Form Converter - Python Dependencies
# Install with: pip install -r requirements.txt

# Requires Python 3.9+

# AWS Integration
boto3==1.34.34
botocore==1.34.34

# Anthropic Claude API
anthropic==0.18.1

# PDF Processing
pdfrw==0.4
pypdf==4.0.1
reportlab==4.1.0

# Image Processing
pillow==10.2.0
pdf2image==1.17.0
numpy==1.26.4

# System Requirements:
# - poppler-utils (for pdf2image)
#   macOS: brew install poppler
#   Ubuntu/Debian: sudo apt-get install poppler-utils
#   Windows: Download from https://github.com/oschwartz10612/poppler-windows/releases/
